requirements.txt
searchtweets==1.7.4
python-dotenv==0.9.1
requests==2.20.1
pystache==0.5.4
feedparser==5.2.1